test(checkbox): migrate color
