+++ core/src/components/checkbox/test/color/checkbox.e2e.ts-snapshots/checkbox-color-checked-ios-ltr-Mobile-Firefox-linux.png
