+++ core/src/components/checkbox/test/color/checkbox.e2e.ts-snapshots/checkbox-color-checked-md-ltr-Mobile-Firefox-linux.png
