+++ core/src/components/checkbox/test/color/checkbox.e2e.ts-snapshots/checkbox-color-unchecked-ios-ltr-Mobile-Safari-linux.png
