+++ core/src/components/checkbox/test/color/checkbox.e2e.ts-snapshots/checkbox-color-unchecked-md-ltr-Mobile-Chrome-linux.png
